backend/requirements.txt
# Kuiper TTS Backend Requirements

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
supabase>=2.0.0

# Utilities
pydantic>=2.0.0
PyJWT[crypto]>=2.8.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
